strip another 1K from libxsm3

+++ lib/xsm3/excrypt_des.h
@@ -9,7 +9,7 @@ typedef struct _EXCRYPT_DES_STATE
 
 void ExCryptDesParity(const uint8_t* input, uint32_t input_size, uint8_t* output);
 
-void ExCryptDesKey(EXCRYPT_DES_STATE* state, const uint8_t* key);
+void ExCryptDesKey(EXCRYPT_DES_STATE* state, const uint64_t* key);
 void ExCryptDesEcb(const EXCRYPT_DES_STATE* state, const uint8_t* input, uint8_t* output, uint8_t encrypt);
 
 typedef struct _EXCRYPT_DES3_STATE
